Create PLSCODE.m
This is a MATLAB script that produces a PLSCODE. The input is the MODCOD  and TYPE fields. The output is the Physical Layer Signalling Code (PLSCODE). 

The PLSCODE and the Start of Frame (SOF) together form the Physical Layer Header (PLHEADER) for each frame in DVB-S2.

The PLHEADER is used by the correlator to obtain frame timing.